requirements.txt
google-genai
watchdog
pydub
python-dotenv
